requirements.txt
aioredis==1.3.1
anyio==3.5.0
asgiref==3.5.0
async-timeout==4.0.2
attrs==21.4.0
autobahn==22.3.2
Automat==20.2.0
cffi==1.15.0
channels==3.0.4
channels-redis==3.4.0
click==8.1.2
constantly==15.1.0
cryptography==36.0.2
daphne==3.0.2
dj-database-url==0.5.0
Django==4.0.3
django-cors-headers==3.11.0
django-filter==21.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
gunicorn==20.1.0
h11==0.13.0
hiredis==2.0.0
httptools==0.4.0
hyperlink==21.0.0
idna==3.3
image==1.5.33
incremental==21.3.0
jwt==1.3.1
msgpack==1.0.3
Pillow==9.1.0
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.3.0
pyOpenSSL==22.0.0
python-dotenv==0.20.0
pytz==2022.1
PyYAML==6.0
service-identity==21.1.0
six==1.16.0
sniffio==1.2.0
sqlparse==0.4.2
Twisted==22.2.0
txaio==22.2.1
typing_extensions==4.1.1
uvicorn==0.17.6
uvloop==0.16.0
watchgod==0.8.2
websockets==10.2
zope.interface==5.4.0
